First attempt.
McV will build and begin to run, but there is apparently some problem
with Java3D and/or JOGL not being able to create the initial display.

I think it may have something to do with my lumping what would
"normally" be local-idv.jar into mcidasv.jar. I seem to remember
something like this happening in the past, but I may also be recalling
things incorrectly!